requirements.txt
asgiref==3.4.1
black==21.12b0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
dj-database-url==2.1.0
Django==4.0
django-heroku==0.3.1
djangorestframework==3.13.1
google-api-core==2.3.2
google-api-python-client==2.33.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
gunicorn==21.2.0
httplib2==0.20.2
idna==3.3
introcs==1.3.1
mypy-extensions==0.4.3
numpy==1.26.4
oauthlib==3.1.1
packaging==24.0
pathspec==0.9.0
pillow==10.2.0
platformdirs==2.4.0
protobuf==3.19.1
psycopg2==2.9.9
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
six==1.16.0
sqlparse==0.4.2
tomli==1.2.3
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==6.6.0
